refactor: refactor InteractionClaim to a common struct for all components
