+++ crates/brainfuck_prover/src/components/processor/component.rs
@@ -2,10 +2,7 @@ use crate::components::{
     instruction::table::InstructionElements, io::table::IoElements, memory::table::MemoryElements,
     ProcessorClaim,
 };
-use stwo_prover::{
-    constraint_framework::{EvalAtRow, FrameworkComponent, FrameworkEval},
-    core::{channel::Channel, fields::qm31::SecureField},
-};
+use stwo_prover::constraint_framework::{EvalAtRow, FrameworkComponent, FrameworkEval};
 
 /// Implementation of `Component` and `ComponentProver`
 /// for the `SimdBackend` from the constraint framework provided by Stwo
@@ -69,27 +66,5 @@ impl FrameworkEval for ProcessorEval {
     /// The logUp must be finalized with `eval.finalize_logup()`.
     fn evaluate<E: EvalAtRow>(&self, _eval: E) -> E {
         todo!()
-    }
-}
-
-/// The claim of the interaction phase 2 (with the logUp protocol).
-///
-/// The total sum is the computed sum of the logUp extension column,
-/// including the padded rows.
-/// It allows proving that the Processor main trace is:
-/// - A permutation of the Memory trace.
-/// - A permutation of a subset of the Instruction trace.
-/// - That the I/O values are the ones in the I/O components, in the same order.
-#[derive(Debug, Eq, PartialEq)]
-pub struct InteractionClaim {
-    /// The computed sum of the logUp extension column, including padded rows.
-    pub claimed_sum: SecureField,
-}
-
-impl InteractionClaim {
-    /// Mix the sums from the logUp protocol into the Fiat-Shamir [`Channel`],
-    /// to bound the proof to the trace.
-    pub fn mix_into(&self, channel: &mut impl Channel) {
-        channel.mix_felts(&[self.claimed_sum]);
     }
 }
